Refactor persistence actors
