+++ api/src/main/scala/hmda/api/http/InstitutionsHttpApi.scala
@@ -15,15 +15,15 @@ import akka.http.scaladsl.model._
 import akka.stream.scaladsl.{ Framing, Sink }
 import akka.util.{ ByteString, Timeout }
 import hmda.api.model._
-import hmda.persistence.{ FilingPersistence, HmdaFileUpload, SubmissionPersistence }
-import hmda.persistence.FilingPersistence.GetFilingByPeriod
-import hmda.persistence.HmdaFileUpload.{ AddLine, _ }
-import hmda.persistence.InstitutionPersistence.GetInstitutionById
-import hmda.persistence.SubmissionPersistence.{ CreateSubmission, GetLatestSubmission }
+import hmda.persistence.institutions.FilingPersistence.GetFilingByPeriod
+import hmda.persistence.institutions.InstitutionPersistence.GetInstitutionById
+import hmda.persistence.institutions.SubmissionPersistence.{ CreateSubmission, GetLatestSubmission }
 import hmda.api.protocol.processing.{ FilingProtocol, InstitutionProtocol }
 import hmda.model.fi.{ Filing, Institution, Submission }
 import hmda.persistence.CommonMessages._
-import hmda.persistence.{ CommonMessages, FilingPersistence, SubmissionPersistence }
+import hmda.persistence.institutions.{ FilingPersistence, SubmissionPersistence }
+import hmda.persistence.CommonMessages
+import hmda.persistence.processing.HmdaFileUpload._
 
 import scala.concurrent.{ ExecutionContext, Future }
 import scala.util.{ Failure, Success }
